+++ docs/source/changelog.rst
@@ -9,6 +9,8 @@ Current
 New Features
 ^^^^^^^^^^^^
 - Add :class:`deepinv.models.DEAL` model (:gh:`939` by `Hossein Alimohammadi`_)
+- Add :class:`deepinv.optim.SIRT` algorithm for tomographic reconstruction (:gh:`985` by `Thibaut Modrzyk`_`)
+- Add :class:`deepinv.optim.MLEM` algorithm for Poisson inverse problems (:gh:`1051` by `Thibaut Modrzyk`_)
 
 Changed
 ^^^^^^^